rust: restrict mobility related code to mobility feature
